Added transaction classes. Other fixes
Read tittle.

+++ src/com/example/a2_4hc2/Withdraw.java
@@ -96,7 +96,7 @@ public class Withdraw extends Activity {
 			
 	         @Override
 	         public void onClick(View v) {
-	        	 int amountToWithdraw = Integer.parseInt(withdrawLabel.getText().toString().substring(1));
+	        	 Double amountToWithdraw = Double.parseDouble((withdrawLabel.getText().toString().substring(1)));
 	        	 
 	        	 RadioButton chequing = (RadioButton) findViewById(R.id.chequing);
